# openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/files/upload:
        post:
            tags:
                - FileService
            operationId: FileService_Upload
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/file.v1.UploadRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/file.v1.UploadReply'
...
components:
    schemas:
        file.v1.UploadReply:
            type: object
            properties:
                name:
                    type: string
                url:
                    type: string
        file.v1.UploadRequest:
            type: object
            properties: {}
...